refactor: move agent definitions to per-file modules
Each agent is now a self-contained file in src/agents/definitions/
implementing AgentDefinition. To add a new agent, create one file
and add a single import to registry.ts.

- claude.ts, cursor.ts, codex.ts, vscode.ts, opencode.ts
- helpers.ts for shared envRecord/httpServer utilities
- registry.ts reduced to imports + Map construction
